# backend/docs/swagger.yaml
basePath: /api/v2
definitions:
  github_com_norman6464_FreStyle_backend_internal_domain.AdminInvitation:
    properties:
      companyId:
        type: integer
      createdAt:
        type: string
      displayName:
        type: string
      email:
        type: string
      expiresAt:
        type: string
      id:
        type: integer
      role:
        type: string
      status:
        type: string
    type: object
...
  github_com_norman6464_FreStyle_backend_internal_usecase_repository.AiChatAttachmentUploadURL:
    properties:
      expiresIn:
        type: integer
      key:
        type: string
      uploadUrl:
        type: string
    type: object
...
  internal_handler.aiChatAttachmentUploadURLRequest:
    properties:
      contentType:
        type: string
      filename:
        type: string
      sizeBytes:
        type: integer
    type: object
...
  internal_handler.createAdminInvReq:
    properties:
      companyId:
        type: integer
      displayName:
        type: string
      email:
        type: string
      role:
        type: string
    required:
    - companyId
    - email
    - role
    type: object
...
  internal_handler.errorResponse:
    properties:
      error:
        example: unauthorized
        type: string
    type: object
...
info:
  contact:
    name: FreStyle Engineering
    url: https://normanblog.com
  description: |-
    新卒 IT エンジニア 向け 統合 研修 プラットフォーム の REST API。
    Clean Architecture (port = usecase/repository / adapter = adapter/persistence) で 構築。
  termsOfService: https://normanblog.com/terms
  title: FreStyle Backend API
  version: "2.0"
paths:
...
  /admin/invitations:
    get:
      description: pending な 招待 を 返す。 SuperAdmin は 全社 (?companyId= で 絞り込み 可)、 CompanyAdmin
        は 自社 のみ。 trainee 等 は 403。
      parameters:
      - description: 'SuperAdmin の とき のみ 有効: 特定 company の 招待 のみ'
        in: query
        name: companyId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_norman6464_FreStyle_backend_internal_domain.AdminInvitation'
            type: array
        "400":
          description: ListByCompanyID 失敗 (現状 実装 で 400 を 返す パス あり)
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: 未 認証
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: trainee / company 未 設定 等
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: DB 失敗 (ListAll 経路)
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - CookieAuth: []
      summary: 招待 一覧 (admin)
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: 'SES マジック リンク で 招待 メール を 送る。 SoD: SuperAdmin は company_admin のみ
        招待 可、 CompanyAdmin は trainee のみ 自社 に 招待 可。'
      parameters:
      - description: 招待 内容 (CompanyAdmin は companyId が 上書き さ れる)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createAdminInvReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_norman6464_FreStyle_backend_internal_domain.AdminInvitation'
        "400":
          description: バリデーション
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: 未 認証
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: ロール 違反
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - CookieAuth: []
      summary: 招待 作成
      tags:
      - admin
  /admin/invitations/{id}:
    delete:
      description: 指定 招待 の status を canceled に 更新。 行 は 物理 削除 せず 監査 用 に 残す。
      parameters:
      - description: 招待 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: 成功 (本文 なし)
        "400":
          description: DB 失敗
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - CookieAuth: []
      summary: 招待 取り消し
      tags:
      - admin
  /ai-chat/attachments/upload-url:
    post:
      consumes:
      - application/json
      description: ai-chat/{userId}/{uuid}.{ext} の キー で S3 PUT 用 presigned URL を 発行。
        contentType は image/png 等 の 許容 セット のみ。 sizeBytes 上限 (image 5MB / document
        4.5MB) も 事前 検証。
      parameters:
      - description: filename / contentType / sizeBytes
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.aiChatAttachmentUploadURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_norman6464_FreStyle_backend_internal_usecase_repository.AiChatAttachmentUploadURL'
        "400":
          description: バリデーション 失敗
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: 未 認証
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "413":
          description: サイズ 上限 超過
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "415":
          description: 未 サポート MIME
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: S3 presigner 失敗
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "503":
          description: 添付 アップロード 機能 が 設定 されて いない (dev/stub)
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - CookieAuth: []
      summary: AI チャット 添付 PUT 署名 URL
      tags:
      - ai-chat
...
securityDefinitions:
  CookieAuth:
    description: Cognito 由来 の JWT を HttpOnly Cookie で 送る。 ログイン 後 自動 付与。
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"